requirements.txt

websocket-client>=1.0.0
Pillow>=9.0.0